+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [0.0.2] - 2025-01-01
 
 Add missing basic features and more integration testing.
 
@@ -26,6 +26,8 @@ Add missing basic features and more integration testing.
 - improve CORS middleware
 - improve authentication middleware
 - improve Node.js facade
+
+[0.0.2]: https://github.com/edadma/apion/releases/tag/v0.0.2
 
 ## [0.0.1] - 2024-12-30